move ObjectStoreType to utils
