+++ slatedb/src/db/builder.rs
@@ -154,10 +154,10 @@ use crate::manifest::store::{ManifestStore, StoredManifest};
 use crate::manifest::ManifestCore;
 use crate::memtable_flusher::MemtableFlusher;
 use crate::merge_operator::MergeOperatorType;
-use crate::object_stores::ObjectStoreType;
 use crate::paths::PathResolver;
 use crate::retrying_object_store::RetryingObjectStore;
 use crate::tablestore::{TableStore, TableStoreKind};
+use crate::utils::ObjectStoreType;
 use crate::utils::SafeSender;
 use crate::utils::WatchableOnceCell;
 use crate::wal;
