+++ slatedb/src/lib.rs
@@ -146,7 +146,6 @@ mod mem_table;
 mod memtable_flusher;
 mod merge_iterator;
 mod merge_operator;
-mod object_stores;
 mod ops;
 mod oracle;
 mod partitioned_keyspace;
