+++ slatedb/src/utils.rs
@@ -27,6 +27,24 @@ use std::collections::VecDeque;
 
 static EMPTY_KEY: Bytes = Bytes::new();
 
+/// Whether the object store holds the main data path or the WAL.
+#[derive(Debug, Clone, Copy)]
+pub(crate) enum ObjectStoreType {
+    /// The primary object store (SSTs, manifests, compacted data).
+    Main,
+    /// The dedicated WAL object store, when configured separately.
+    Wal,
+}
+
+impl ObjectStoreType {
+    pub(crate) fn as_str(self) -> &'static str {
+        match self {
+            Self::Main => "main",
+            Self::Wal => "wal",
+        }
+    }
+}
+
 #[derive(Clone, Debug)]
 pub(crate) struct WatchableOnceCell<T: Clone> {
     rx: tokio::sync::watch::Receiver<Option<T>>,
